agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helios-agent"
version = "0.1.0"
description = "Metrics collection agent for Helios - Predictive Infrastructure Intelligence Platform"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "Helios Platform", email = "maintainers@helios.dev"}
]
keywords = [
    "kubernetes",
    "metrics",
    "monitoring",
    "prometheus",
    "observability"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "httpx>=0.24.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
prometheus = [
    "prometheus-client>=0.17.0",
]
datadog = [
    "datadog-api-client>=2.0.0",
]
aws = [
    "boto3>=1.26.0",
]
azure = [
    "azure-identity>=1.12.0",
    "azure-mgmt-monitor>=6.0.0",
]
gcp = [
    "google-cloud-monitoring>=2.15.0",
]
kubernetes = [
    "kubernetes>=28.0.0",
]
all = [
    "helios-agent[prometheus,datadog,aws,azure,gcp,kubernetes]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
helios-agent = "helios_agent.cli:main"

[project.urls]
Homepage = "https://github.com/pyjeebz/helios"
Repository = "https://github.com/pyjeebz/helios"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
